reorganizing the famous radiation.py into it's own package
